Added some basic file support for inmemory db

+++ server/src/workflows/entity.rs
@@ -4,11 +4,10 @@ use chrono::Utc;
 use chrono::DateTime;
 use database::traits::Entity;
 
-
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct Workflow {
-    #[serde(skip)]
-    pub id: i32,
+    #[serde(default)]
+    pub id: u64,
 
     pub name: String,
 
@@ -23,6 +22,6 @@ pub struct Workflow {
 }
 
 impl Entity for Workflow {
-    fn get_id(&self) -> i32 { self.id }
-    fn set_id(&mut self, id: i32) { self.id = id; }
+    fn get_id(&self) -> u64 { self.id }
+    fn set_id(&mut self, id: u64) { self.id = id; }
 }
